Renamed to changeNotificationPreference
