+++ auth-server/src/resolvers/index.ts
@@ -3,6 +3,7 @@ import addressLinkStart from './mutation/addressLinkStart';
 import addressUnlink from './mutation/addressUnlink';
 import changeUsername from './mutation/changeUsername';
 import changeEmail from './mutation/changeEmail';
+import changeNotificationPreference from './mutation/changeNotificationPreference';
 import changePassword from './mutation/changePassword';
 import changeName from './mutation/changeName';
 import login from './mutation/login';
@@ -14,7 +15,6 @@ import resendVerifyEmailToken from './mutation/resendVerifyEmailToken';
 import resetPassword from './mutation/resetPassword';
 import signup from './mutation/signup';
 import undoEmailChange from './mutation/undoEmailChange';
-import updateNotification from './mutation/updateNotification';
 import verifyEmail from './mutation/verifyEmail';
 import subscription from './query/subscription';
 import token from './query/token';
@@ -37,6 +37,7 @@ export default {
 		signup,
 		changeUsername,
 		changeEmail,
+		changeNotificationPreference,
 		changePassword,
 		changeName,
 		postSubscribe,
@@ -45,7 +46,6 @@ export default {
 		resendVerifyEmailToken,
 		resetPassword,
 		undoEmailChange,
-		updateNotification,
 		verifyEmail
 	}
 };
