+++ auth-server/src/resolvers/mutation/changeNotificationPreference.ts
