+++ auth-server/src/schema/mutation/changeNotificationPreference.ts
@@ -1,5 +1,5 @@
 export default `
-    updateNotification(
+    changeNotificationPreference(
         post_participated: Boolean
         post_created: Boolean
         new_proposal: Boolean
